+++ cookbooks/curator/attributes/default.rb
@@ -1,0 +1,3 @@
+# Attributes for Curator
+
+default["curator"]["dir"] = "/home/#{node["delk_user"]}/curator"