+++ cookbooks/kibana/attributes/default.rb
@@ -1,0 +1,4 @@
+# Attributes for Kibana
+
+default["kibana"]["dir"]   = "/home/#{node["delk_user"]}/kibana"
+default["kibana"]["port"]  = "5601"