+++ cookbooks/logstash/attributes/default.rb
@@ -1,4 +1,4 @@
 # Attributes for Logstash
 
-default["logstash"]["dir"]    = "/home/vagrant/logstash"
+default["logstash"]["dir"]    = "/home/#{node["delk_user"]}/logstash"
 default["logstash"]["port"]   = "5000"