+++ cookbooks/post-setup/metadata.rb
@@ -1,0 +1,6 @@
+name    'post-setup'
+version '0.0.1'
+
+depends 'elasticsearch', '>=0.0.1'
+depends 'logstash', '>=0.0.1'
+depends 'kibana', '>=0.0.1'